requirements.txt
numpy==1.24.3
Pillow==10.0.1
boto3==1.34.0
requests
pyyaml